Move `sr-pipeline` script into `sr-stats`
