--- libs/sr-pipeline/README.md
@@ -1,11 +1,0 @@
-# sr-pipeline
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build sr-pipeline` to build the library.
-
-## Running unit tests
-
-Run `nx test sr-pipeline` to execute the unit tests via [Jest](https://jestjs.io).
