--- libs/sr-pipeline/jest.config.ts
@@ -1,15 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'sr-pipeline',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/libs/sr-pipeline',
-}
