--- libs/sr-pipeline/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@genshin-optimizer/sr-pipeline",
-  "version": "0.0.1",
-  "type": "commonjs"
-}
