--- libs/sr-pipeline/project.json
@@ -1,31 +1,0 @@
-{
-  "name": "sr-pipeline",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/sr-pipeline/src",
-  "projectType": "library",
-  "targets": {
-    "build": {},
-    "build-ts": {
-      "options": {
-        "outputFileName": "src/index.js"
-      }
-    },
-    "pipeline": {
-      "executor": "@nx/js:node",
-      "options": {
-        "buildTarget": "sr-pipeline:build-ts",
-        "watch": false,
-        "inspect": false
-      },
-      "inputs": ["default", "^production", "sr-stats"]
-    },
-    "lint": {}
-  },
-  "tags": [],
-  "namedInputs": {
-    "sr-stats": [
-      "{workspaceRoot}/lib/sr-stats/Data/**/*.json",
-      "{workspaceRoot}/lib/sr-stats/src/allStat_gen.json"
-    ]
-  }
-}
