+++ libs/sr-stats/project.json
@@ -4,12 +4,16 @@
   "sourceRoot": "libs/sr-stats/src",
   "projectType": "library",
   "targets": {
-    "build": {},
-    "build-ts": {
-      "dependsOn": ["^build", "^pipeline"]
+    "genfile": {
+      "executor": "@genshin-optimizer/sr-stats:gen-stats",
+      "outputs": [
+        "{projectRoot}/Data/**/*",
+        "{projectRoot}/src/allStat_gen.json"
+      ]
     },
+    "build": {},
+    "build-ts": {},
     "lint": {}
   },
-  "tags": [],
-  "implicitDependencies": ["sr-pipeline"]
+  "tags": []
 }
