+++ libs/sr-stats/src/executors/gen-stats/schema.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "version": 2,
+  "title": "GenStats executor",
+  "description": "",
+  "type": "object",
+  "properties": {},
+  "required": []
+}
