+++ libs/sr-stats/src/executors/gen-stats/src/characterData.ts
