+++ libs/sr-stats/src/executors/gen-stats/src/lightConeData.ts
