+++ libs/sr-stats/src/index.ts
@@ -1,5 +1,11 @@
-import type { AllStats } from '@genshin-optimizer/sr-pipeline'
 import * as allStat_gen from './allStat_gen.json'
+
+// Make sure these are type-only imports/exports.
+// Importing the executor is quite costly.
+export type {} from './executors/gen-stats/executor'
+import type { AllStats } from './executors/gen-stats/executor'
+
 const allStats = allStat_gen as AllStats
 
+export type { AllStats }
 export { allStats }
