+++ tsconfig.base.json
@@ -48,7 +48,6 @@
       "@genshin-optimizer/sr-consts": ["libs/sr-consts/src/index.ts"],
       "@genshin-optimizer/sr-dm": ["libs/sr-dm/src/index.ts"],
       "@genshin-optimizer/sr-formula": ["libs/sr-formula/src/index.ts"],
-      "@genshin-optimizer/sr-pipeline": ["libs/sr-pipeline/src/index.ts"],
       "@genshin-optimizer/sr-stats": ["libs/sr-stats/src/index.ts"],
       "@genshin-optimizer/svgicons": ["libs/svgicons/src/index.ts"],
       "@genshin-optimizer/ui-common": ["libs/ui-common/src/index.ts"],
